requirements.txt
alembic==1.10.2
async-timeout==4.0.2
blinker==1.5
cachelib==0.10.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-JWT==0.3.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.0
Pillow==9.4.0
PyJWT==1.4.2
python-dotenv==1.0.0
python-magic==0.4.27
redis==4.5.1
requests==2.31.0
SQLAlchemy==2.0.6
SQLAlchemy-ImageAttach==1.1.0
stripe==6.3.0
typing_extensions==4.5.0
urllib3==2.0.4
Wand==0.6.11
Werkzeug==2.2.3
